Rename auth options
